add validation

+++ src/main/java/org/boomzin/votehub/web/menu/MenuController.java
@@ -2,6 +2,7 @@ package org.boomzin.votehub.web.menu;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.boomzin.votehub.error.IllegalRequestDataException;
 import org.boomzin.votehub.model.MenuItem;
 import org.boomzin.votehub.model.Restaurant;
 import org.boomzin.votehub.repository.MenuItemRepository;
@@ -10,17 +11,22 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import javax.validation.Valid;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static org.boomzin.votehub.util.ValidationUtil.assureIdConsistent;
 import static org.boomzin.votehub.util.ValidationUtil.checkNew;
 
+@Validated
 @RestController
 @RequestMapping(value = MenuController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
 @Slf4j
@@ -33,10 +39,14 @@ public class MenuController {
 
     @Transactional
     @PostMapping(value = "/{restaurantId}/menu", consumes = MediaType.APPLICATION_JSON_VALUE)
+// TODO: validate collection doesn't work
     public ResponseEntity<List<MenuItem>> createMenuWithLocation(@PathVariable int restaurantId, @RequestBody List<MenuItem> menu) {
         log.info("create menu for restaurant {} for today", restaurantId);
+        Set<String> descriptions = menu.stream().map(x -> x.getDescription().toLowerCase()).collect(Collectors.toSet());
+        if (descriptions.size() != menu.size()) {
+            throw new IllegalRequestDataException("Menu contains not unique dishes");
+        }
         menuItemRepository.deleteAllActualItemsForRestaurant(restaurantId);
-        // TODO: add description_date unique validator
         Restaurant restaurant = restaurantRepository.getById(restaurantId);
         for (MenuItem menuItem : menu) {
             checkNew(menuItem);
@@ -68,22 +78,31 @@ public class MenuController {
     @Transactional
     @PutMapping(value = "/{restaurantId}/menu/menuitem/{menuItemId}", consumes = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void UpdateMenuItem(@PathVariable int restaurantId, @PathVariable int menuItemId, @RequestBody MenuItem menuItem) {
+    public void UpdateMenuItem(@PathVariable int restaurantId, @PathVariable int menuItemId,@Valid @RequestBody MenuItem menuItem) {
         log.info("update menu item {} for restaurant {} for today",menuItemId, restaurantId);
         assureIdConsistent(menuItem, menuItemId);
         Optional<Restaurant> restaurant = restaurantRepository.getWithActualMenu(restaurantId);
         if (restaurant.isPresent() && restaurant.get().getMenu().contains(menuItem)) {
             menuItem.setDate(LocalDate.now());
-            menuItem.setDescription(menuItem.getDescription().toLowerCase());
+            MenuItem existed = menuItemRepository.getById(menuItemId);
+            menuItem.setDescription(existed.getDescription());
             menuItem.setRestaurant(restaurant.get());
             menuItemRepository.save(menuItem);
+        } else {
+            throw new IllegalRequestDataException("Restaurant menu for today doesn`t contain this dish, use POST to add new");
         }
     }
 
     @Transactional
     @PostMapping(value = "/{restaurantId}/menu/menuitem", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<MenuItem> createMenuItemWithLocation(@PathVariable int restaurantId, @RequestBody MenuItem menuItem) {
+    public ResponseEntity<MenuItem> createMenuItemWithLocation(@PathVariable int restaurantId, @Valid @RequestBody MenuItem menuItem) {
         log.info("create menu item for restaurant {} for today", restaurantId);
+        List<MenuItem> actualMenu = menuItemRepository.getActualMenuCurrentRestaurant(restaurantId).get();
+        Set<String> descriptions = actualMenu.stream().map(x -> x.getDescription().toLowerCase()).collect(Collectors.toSet());
+        descriptions.add(menuItem.getDescription().toLowerCase());
+        if (descriptions.size() == actualMenu.size()) {
+            throw new IllegalRequestDataException("Menu already contains this item, use method PUT to change price");
+        }
         Restaurant restaurant = restaurantRepository.getById(restaurantId);
         checkNew(menuItem);
         menuItem.setDate(LocalDate.now());
@@ -96,6 +115,8 @@ public class MenuController {
         return ResponseEntity.created(uriOfNewResource).body(created);
     }
 
+
+
     @GetMapping("/{restaurantId}/menu/menuitem/{menuItemId}")
     public MenuItem getMenuItem (@PathVariable int restaurantId, @PathVariable int menuItemId) {
         log.info("get menuItem {} from actual menu for restaurant {}", menuItemId, restaurantId);
@@ -106,6 +127,7 @@ public class MenuController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteMenuItemForCurrentRestaurant(@PathVariable int restaurantId, @PathVariable int menuItemId) {
         log.info("delete menuItem {} from actual menu for restaurant {}", menuItemId, restaurantId);
-        menuItemRepository.deleteActualItemForRestaurant(restaurantId, menuItemId);
+        menuItemRepository.checkBelong(restaurantId, menuItemId);
+        menuItemRepository.delete(menuItemId);
     }
 }
